# feat(strmap): new pkg for optimizing GC issues