--- a/container/strmap/utils_test.go
+++ b/container/strmap/utils_test.go
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2024 CloudWeGo Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package strmap
+
+import (
+	"testing"
+)
+
+func TestCalcHashtableSlots(t *testing.T) {
+	n := 0
+	for n < 2147483647 {
+		m := calcHashtableSlots(n)
+		t.Log(n, m)
+		n = int(m) + 1
+	}
+}