… hash/xfnv/xfnv.go
… hash/xfnv/xfnv_test.go